requirements.txt
matplotlib
python-telegram-bot == 13.15